pyproject.toml
[project]
name = "video-to-3dgs"
version = "0.1.0"
description = "360度動画から3D Gaussian Splatting用データセット生成ツール"
requires-python = "~=3.12.0"
dependencies = [
    # GUI
    "customtkinter>=5.2.2",
    
    # Computer Vision  
    "opencv-python>=4.8.0",
    "ultralytics>=8.0.0",
    
    # CUDA/GPU
    "cupy-cuda12x>=12.0.0", 
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    
    # 数値計算・画像処理
    "numpy>=1.24.0",
    "pillow>=10.0.0", 
    "scipy>=1.11.0",
    
    # 3D処理
    "open3d>=0.17.0",
    "trimesh>=3.20.0",
    
    # データ処理
    "pandas>=2.0.0",
    "h5py>=3.9.0",
    
    # システム・並行処理
    "psutil>=5.9.0",
    "tqdm>=4.65.0",
    
    # 設定・ログ
    "pyyaml>=6.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0", 
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.hatch.build.targets.wheel]
packages = [
  "core",
  "utils",
  "models",
  "scripts"
]